backend/known_fakes/requirements.txt
flask
python-dotenv
request
